refactor: separated app logic into its own module
